+++ lib/host_kit/remote.ex
@@ -9,6 +9,7 @@ defmodule HostKit.Remote do
     Command,
     Directory,
     EnvFile,
+    Exs,
     File,
     Mise,
     Package,
@@ -88,6 +89,10 @@ defmodule HostKit.Remote do
 
   def read(%HostKit.Resources.Template{} = desired, context) do
     Helpers.read_template(desired, &read(&1, context))
+  end
+
+  def read(%Exs{} = desired, context) do
+    Helpers.read_exs(desired, &read(&1, context))
   end
 
   def read(%Caddy.Site{} = desired, context) do